pyproject.toml
[tool.poetry]
name = "roche"
version = "0.1.0"
description = ""
authors = ["gioaca00 <giovanniacampa00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymed = "^0.8.9"
biopython = "^1.84"
questionary = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
